// src/app/(main)/(pages)/logs/data/schema.ts
import { z } from "zod"

// We're keeping a simple non-relational schema here.
// IRL, you will have a schema for your data models.
export const taskSchema = z.object({
  id: z.string(),
  title: z.string(),
  status: z.string(),
  label: z.string(),
  dateHour: z.string(),
})

export type Task = z.infer<typeof taskSchema>